[#17] Refactor code and test to be more reusable
