+++ spec/requests/api/v1/tokens_controller_spec.rb
@@ -2,11 +2,11 @@
 
 require 'rails_helper'
 
-describe Doorkeeper::TokensController, type: :request do
+describe API::V1::TokensController, type: :request do
   context 'when a user asks for a token with good request' do
     it 'returns a refreshable token' do
       post :create, params: token_request_params
-      expect(JSON.parse(response.body)['token_type']).to eq('Bearer')
+      expect(JSON.parse(response.body)['data']['attributes']['token_type']).to eq('Bearer')
     end
   end
 
@@ -16,7 +16,7 @@ describe Doorkeeper::TokensController, type: :request do
 
       post :create, params: params
 
-      expect(JSON.parse(response.body)['error']).to eq('invalid_client')
+      expect(JSON.parse(response.body)['errors'][0]['code']).to eq('invalid_client')
     end
   end
 
@@ -26,7 +26,7 @@ describe Doorkeeper::TokensController, type: :request do
 
       post :create, params: params
 
-      expect(JSON.parse(response.body)['error']).to eq('invalid_grant')
+      expect(JSON.parse(response.body)['errors'][0]['code']).to eq('invalid_grant')
     end
   end
 
@@ -35,18 +35,18 @@ describe Doorkeeper::TokensController, type: :request do
       token = query_token
       refreshed_token = query_refresh_token token['refresh_token']
 
-      expect(refreshed_token['token_type']).to eq('Bearer')
+      expect(refreshed_token['data']['attributes']['token_type']).to eq('Bearer')
     end
   end
 
   context 'when a user revokes an existing token' do
-    it 'returns no error message' do
+    it 'returns a success status code' do
       token = query_token
       params = token_request_params.merge({ token: token })
 
       post :revoke, params: params
 
-      expect(JSON.parse(response.body)).to eq({})
+      expect(response.status).to eq(200)
     end
   end
 end
